+++ MozoSDK/Classes/ModuleDependencies.swift
@@ -549,10 +549,10 @@ class ModuleDependencies {
     func getConversationList(page: Int, size: Int) -> Promise<[Conversation]> {
        return (coreWireframe.corePresenter?.coreInteractorService?.getConversationList(page: page, size: size))!
     }
-    func getChatMessages(id: Int, page: Int, size: Int) -> Promise<[ConversationMessage]> {
+    func getChatMessages(id: Int64, page: Int, size: Int) -> Promise<[ConversationMessage]> {
        return (coreWireframe.corePresenter?.coreInteractorService?.getChatMessages(id: id, page: page, size: size))!
     }
-    func updateReadConversation(conversationId: Int, lastMessageId: Int) -> Promise<Any> {
+    func updateReadConversation(conversationId: Int64, lastMessageId: Int) -> Promise<Any> {
         return (coreWireframe.corePresenter?.coreInteractorService?.updateReadConversation(conversationId: conversationId, lastMessageId: lastMessageId))!
     }
