+++ MozoSDK/Classes/Modules/Core/Business Logic/Interactor/CoreInteractorIO.swift
@@ -210,6 +210,6 @@ protocol CoreInteractorService {
     
     // MARK: Mozo Messages APIs
     func getConversationList(page: Int, size: Int) -> Promise<[Conversation]>
-    func getChatMessages(id: Int, page: Int, size: Int) -> Promise<[ConversationMessage]>
-    func updateReadConversation(conversationId: Int, lastMessageId: Int) -> Promise<Any>
+    func getChatMessages(id: Int64, page: Int, size: Int) -> Promise<[ConversationMessage]>
+    func updateReadConversation(conversationId: Int64, lastMessageId: Int) -> Promise<Any>
 }
